# Arm64 atomic: add LSE atomics instructions to arm64