@lubinszARM 2022-04-14
After comparing with similar code in Golang, I thinks it's OK.
LGTM.